pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["Julia Mezenova <julia21moon86@mail.ru>"]
readme = "README.md"
packages = [{include = "task_manager"}]

[tool.poetry.dependencies]
python = "^3.8"
psycopg2-binary = "^2.9.9"
dj-database-url = "^2.1.0"
whitenoise = {extras = ["brotli"], version = "^6.6.0"}
python-dotenv = "^1.0.1"
gunicorn = "^21.2.0"
uvicorn = "^0.27.1"
django-bootstrap5 = "^23.3"
coverage = "^7.4.3"
django-filter = "^24.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
flake8 = "^4.0.1"
